moved files to lib directory, only main api entry points remain in root
