+++ create.js
@@ -1,8 +1,8 @@
-var Selector = require('./selector')
-var finders = require('./finders')
-var actions = require('./actions')
-var assertions = require('./assertions')
-var promise = require('./promise')
+var Selector = require('./lib/selector')
+var finders = require('./lib/finders')
+var actions = require('./lib/actions')
+var assertions = require('./lib/assertions')
+var promise = require('./lib/promise')
 
 module.exports = function (rootSelector) {
   return new Selector(rootSelector)
