+++ lib/assertions.js
