+++ lib/elementTester.js
