+++ lib/elementsToString.js
