+++ lib/errorHandler.js
