+++ lib/finders.js
