+++ lib/options.js
