+++ lib/selector.js
