+++ lib/stubBrowser.js
