+++ utils/removeStrictFromSend.js
